+++ config/default.js
@@ -145,6 +145,14 @@ module.exports = {
     "port": process.env.PORT || 60080,
     "auth_strategy": "bearer",
   },
-  "dispatcher": {
+  "events": {
+    "user": {
+      "email" : "support@theeye.io",
+      "username" : "theeye",
+      "enabled" : true,
+      "client_id":null,
+      "client_secret":null,
+      "token":null
+    }
   }
 }