+++ core/controllers/event.js
@@ -4,6 +4,8 @@ var appRoot = require('app-root-path');
 var resolver = require('../router/param-resolver');
 var Event = require('../entity/event').Event;
 
+var async = require('async');
+
 module.exports = function (server, passport) {
   server.get('/:customer/event/:event',[
     passport.authenticate('bearer', {session:false}),
@@ -25,16 +27,23 @@ var controller = {
   fetch (req, res, next) {
     Event
     .find({ customer: req.customer })
-    .populate({
-      path: 'emitter',
-      populate: {
-        path: 'host',
-        model: 'Host'
-      }
-    })
     .exec(function(err, events){
-      if(err) res.send(500);
-      res.send(200, events);
+      // theres is a bug in mongoose with this schemas
+      // populate within the find query does not work as expected
+      async.each(
+        events,
+        (e, done) => e.populate({
+          path: 'emitter',
+          populate: {
+            path: 'host',
+            model: 'Host'
+          }
+        }, done),
+        (err) => {
+          if(err) res.send(500);
+          res.send(200, events);
+        }
+      );
     });
   },
 }