+++ core/service/job.js
@@ -111,7 +111,7 @@ var service = {
       throw new Error('invalid or undefined task type ' + task.type);
     }
   },
-  fetchBy(input,next) {
+  fetchBy (input,next) {
     var query = {};
 
     if( input.host ) query.host_id = input.host._id ;
@@ -174,7 +174,7 @@ var service = {
 
         NotificationService.sendEmailNotification({
           customer_name: job.customer_name,
-          subject: `[Task] ${job.task.name} executed on ${job.host.hostname}`,
+          subject: `[TASK] ${job.task.name} executed on ${job.host.hostname}`,
           content: html,
           to: job.user.email
         });